What's new 1. Changed local url to Dev URL 2. Changed the text size for smaller screens 3. Reduced Button size on Login and forgot password

+++ lib/models/login_flow/login_pending_products_dto.dart
@@ -7,7 +7,7 @@ part 'login_pending_products_dto.g.dart';
 class LoginPendingProductsDTO {
   int id;
   String product;
-  int loanAmount;
+  double loanAmount;
   List<Individual> applicants;
 
   LoginPendingProductsDTO({
